added exceptions to boarding mgt

+++ boarding_mgt/Controllers/BoardingApiController.cs
@@ -1,7 +1,9 @@
 ﻿using boarding_mgt.Data;
 using boarding_mgt.Models;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Connections;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace boarding_mgt.Controllers
 {
@@ -16,22 +18,32 @@ namespace boarding_mgt.Controllers
             _db = db;
         }
         [HttpGet]
-        public ActionResult<IEnumerable<Boarding>> GetBoardings()
+        public async Task< ActionResult<IEnumerable<Boarding>>> GetBoardings()
         {
-            return _db.Boarding;
+            var boardings = await _db.Boarding.ToListAsync();
+            if (boardings == null)
+                return NotFound("There are no boarding available");
+            return Ok(boardings);
         }
 
         [HttpGet("{Id:int}")]
         public async Task<ActionResult<Boarding>> GetById(int Id)
         {
             var boarding = await _db.Boarding.FindAsync(Id);
+            if(boarding==null)
+                return NotFound("Boarding not found");
+
             return boarding;
         }
 
         [HttpPost]
         public async Task<ActionResult> CreateBoarding(Boarding boarding)
         {
-            await _db.AddAsync(boarding);
+            var ExistingBoarding = await _db.Boarding.FirstOrDefaultAsync(b => b.Name == boarding.Name);
+            if (ExistingBoarding != null)
+                return Conflict($"Boarding with name '{boarding.Name}' already exists." );
+        
+        await _db.AddAsync(boarding);
             await _db.SaveChangesAsync();
             return Ok();
         }
@@ -39,7 +51,12 @@ namespace boarding_mgt.Controllers
         [HttpPut]
         public async Task<ActionResult> UpdateBoarding(Boarding boarding)
         {
-            _db.Boarding.Update(boarding);
+            var Boarding = await _db.Boarding.FindAsync(boarding.Id);
+            if (Boarding == null)
+                return NotFound("Boarding not found");
+            Boarding.Name = boarding.Name;
+            Boarding.Description = boarding.Description;
+            Boarding.Type = boarding.Type;
             await _db.SaveChangesAsync();
             return Ok();
         }
@@ -48,6 +65,8 @@ namespace boarding_mgt.Controllers
         public async Task<ActionResult> DeleteBoarding(int Id)
         {
             var temp = await _db.Boarding.FindAsync(Id);
+            if (temp == null)
+                return NotFound("Boarding not found");
             _db.Boarding.Remove(temp);
             await _db.SaveChangesAsync();
             return Ok();
